Fix SIGSEGV in demo program when compiled with gcc (#12)
When compiled with gcc, the big stack test that runs as part of the demo/test
program hits a SIGSEGV when it attempts to access beyond the max fiber stack
size of 64KiB. The test was supposed to perform 15 page size stack allocations,
leaving some space for other stack use added by the compiler, but was mistakenly
implemetned to perform 16 of them. This works when compiled with clang (though
the 17th will hit a SIGSEGV) but it can only allocate 15 when compiled with gcc.
This change fixes the original bug in the test and the test runs fine when
compiled with either gcc or clang.